SHIRO-618 Remove Spring duplicate auto config classes
Including shiro-spring-boot-starter as a shiro-spring-boot-web-starter dependency
