+++ support/spring-boot/spring-boot-starter/src/main/java/org/apache/shiro/spring/boot/autoconfigure/exception/NoRealmBeanConfiguredException.java
@@ -1,0 +1,9 @@
+package org.apache.shiro.spring.boot.autoconfigure.exception;
+
+/**
+ * This exception should be thrown if not bean of type {@link org.apache.shiro.realm.Realm Realm} found.
+ *
+ */
+public class NoRealmBeanConfiguredException extends RuntimeException {
+
+}
