+++ support/spring-boot/spring-boot-web-starter/pom.xml
@@ -35,7 +35,7 @@
 
         <dependency>
             <groupId>org.apache.shiro</groupId>
-            <artifactId>shiro-spring</artifactId>
+            <artifactId>shiro-spring-boot-starter</artifactId>
         </dependency>
 
         <dependency>
