+++ support/spring-boot/spring-boot-web-starter/src/main/java/org/apache/shiro/spring/config/web/autoconfigure/ShiroWebAutoConfiguration.java
@@ -23,23 +23,22 @@ import java.util.List;
 import org.apache.shiro.authc.Authenticator;
 import org.apache.shiro.authc.pam.AuthenticationStrategy;
 import org.apache.shiro.authz.Authorizer;
-import org.apache.shiro.config.Ini;
 import org.apache.shiro.mgt.RememberMeManager;
 import org.apache.shiro.mgt.SessionStorageEvaluator;
 import org.apache.shiro.mgt.SessionsSecurityManager;
 import org.apache.shiro.mgt.SubjectDAO;
 import org.apache.shiro.mgt.SubjectFactory;
 import org.apache.shiro.realm.Realm;
-import org.apache.shiro.realm.text.IniRealm;
 import org.apache.shiro.session.mgt.SessionFactory;
 import org.apache.shiro.session.mgt.SessionManager;
 import org.apache.shiro.session.mgt.eis.SessionDAO;
-import org.apache.shiro.spring.config.web.autoconfigure.exception.NoRealmBeanConfiguredException;
+import org.apache.shiro.spring.boot.autoconfigure.ShiroAutoConfiguration;
 import org.apache.shiro.spring.web.config.AbstractShiroWebConfiguration;
+import org.apache.shiro.spring.web.config.ShiroFilterChainDefinition;
 import org.apache.shiro.web.servlet.Cookie;
+import org.springframework.boot.autoconfigure.AutoConfigureBefore;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnResource;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -47,6 +46,7 @@ import org.springframework.context.annotation.Configuration;
  * @since 1.4.0
  */
 @Configuration
+@AutoConfigureBefore(ShiroAutoConfiguration.class)
 @ConditionalOnProperty(name = "shiro.web.enabled", matchIfMissing = true)
 public class ShiroWebAutoConfiguration extends AbstractShiroWebConfiguration {
 
@@ -88,6 +88,13 @@ public class ShiroWebAutoConfiguration extends AbstractShiroWebConfiguration {
     @Bean
     @ConditionalOnMissingBean
     @Override
+    protected SubjectFactory subjectFactory() {
+        return super.subjectFactory();
+    }
+
+    @Bean
+    @ConditionalOnMissingBean
+    @Override
     protected SessionFactory sessionFactory() {
         return super.sessionFactory();
     }
@@ -97,6 +104,20 @@ public class ShiroWebAutoConfiguration extends AbstractShiroWebConfiguration {
     @Override
     protected SessionDAO sessionDAO() {
         return super.sessionDAO();
+    }
+
+    @Bean
+    @ConditionalOnMissingBean
+    @Override
+    protected SessionManager sessionManager() {
+        return super.sessionManager();
+    }
+
+    @Bean
+    @ConditionalOnMissingBean
+    @Override
+    protected SessionsSecurityManager securityManager(List<Realm> realms) {
+        return super.securityManager(realms);
     }
 
     @Bean
@@ -123,44 +144,7 @@ public class ShiroWebAutoConfiguration extends AbstractShiroWebConfiguration {
     @Bean
     @ConditionalOnMissingBean
     @Override
-    protected SubjectFactory subjectFactory() {
-        return super.subjectFactory();
+    protected ShiroFilterChainDefinition shiroFilterChainDefinition() {
+        return super.shiroFilterChainDefinition();
     }
-
-    @Bean
-    @ConditionalOnMissingBean
-    @Override
-    protected SessionManager sessionManager() {
-        return super.sessionManager();
-    }
-
-    @Bean
-    @ConditionalOnMissingBean
-    @Override
-    protected SessionsSecurityManager securityManager(List<Realm> realms) {
-        return super.securityManager(realms);
-    }
-    
-    @Bean
-    @ConditionalOnResource(resources = "classpath:shiro.ini")
-    protected Realm iniClasspathRealm() {
-        Ini ini = Ini.fromResourcePath("classpath:shiro.ini");
-        IniRealm iniRealm = new IniRealm( ini );
-        return iniRealm;
-    }
-    
-    @Bean
-    @ConditionalOnResource(resources = "classpath:META-INF/shiro.ini")
-    protected Realm iniMetaInfClasspathRealm() {
-        Ini ini = Ini.fromResourcePath("classpath:META-INF/shiro.ini");
-        IniRealm iniRealm = new IniRealm( ini );
-        return iniRealm;
-    }
-    
-    @Bean
-    @ConditionalOnMissingBean(Realm.class)
-    protected Realm missingRealm() {
-        throw new NoRealmBeanConfiguredException();
-    }
-
 }
