--- support/spring-boot/spring-boot-web-starter/src/main/java/org/apache/shiro/spring/config/web/autoconfigure/exception/NoRealmBeanConfiguredException.java
@@ -1,9 +1,0 @@
-package org.apache.shiro.spring.config.web.autoconfigure.exception;
-
-/**
- * This exception should be thrown if not bean of type {@link Realm} found.
- *
- */
-public class NoRealmBeanConfiguredException extends RuntimeException {
-
-}
